requirement.txt
urllib
python-rake
textrazor
rake-nltk
pytextrank